Refactor the internal/template folder and add a proper documentation
