--- internal/templates/router_routes.go
@@ -1,48 +1,0 @@
-package template
-
-type HttpRouterRouteTemplate struct {
-}
-
-func (c HttpRouterRouteTemplate) Main() []byte {
-	return MainTemplate()
-}
-func (c HttpRouterRouteTemplate) Server() []byte {
-	return MakeHTTPServer()
-}
-func (c HttpRouterRouteTemplate) Routes() []byte {
-	return MakeRouterRoutes()
-}
-
-func MakeRouterRoutes() []byte {
-	return []byte(`package server
-
-import (
-	"encoding/json"
-	"log"
-	"net/http"
-
-	"github.com/julienschmidt/httprouter"
-)
-
-func (s *Server) RegisterRoutes() http.Handler {
-	r := httprouter.New()
-	r.HandlerFunc(http.MethodGet, "/", s.helloWorldHandler)
-
-	return r
-}
-
-func (s *Server) helloWorldHandler(w http.ResponseWriter, r *http.Request) {
-	resp := make(map[string]string)
-	resp["message"] = "Hello World"
-
-	jsonResp, err := json.Marshal(resp)
-	if err != nil {
-		log.Fatalf("error handling JSON marshal. Err: %v", err)
-	}
-
-	w.Write(jsonResp)
-}
-
-
-`)
-}
