--- internal/templates/standard_routes.go
@@ -1,80 +1,0 @@
-package template
-
-type StandardLibraryRouteTemplate struct {
-}
-
-func (c StandardLibraryRouteTemplate) Main() []byte {
-	return MainTemplate()
-}
-func (c StandardLibraryRouteTemplate) Server() []byte {
-	return MakeHTTPServer()
-}
-func (c StandardLibraryRouteTemplate) Routes() []byte {
-	return MakeHTTPRoutes()
-}
-
-func MakeHTTPServer() []byte {
-	return []byte(`package server
-
-import (
-	"fmt"
-	"net/http"
-	"time"
-)
-
-var port = 8080
-
-type Server struct {
-	port int
-}
-
-func NewServer() *http.Server {
-
-	NewServer := &Server{
-		port: port,
-	}
-
-	// Declare Server config
-	server := &http.Server{
-		Addr:         fmt.Sprintf(":%d", NewServer.port),
-		Handler:      NewServer.RegisterRoutes(),
-		IdleTimeout:  time.Minute,
-		ReadTimeout:  10 * time.Second,
-		WriteTimeout: 30 * time.Second,
-	}
-
-	return server
-}
-`)
-}
-
-func MakeHTTPRoutes() []byte {
-	return []byte(`package server
-
-import (
-	"net/http"
-	"encoding/json"
-	"log"
-)
-
-func (s *Server) RegisterRoutes() http.Handler {
-
-	mux := http.NewServeMux()
-	mux.HandleFunc("/", s.handler)
-
-	return mux
-}
-
-func (s *Server) handler(w http.ResponseWriter, r *http.Request) {
-	resp := make(map[string]string)
-	resp["message"] = "Hello World"
-
-	jsonResp, err := json.Marshal(resp)
-	if err != nil {
-		log.Fatalf("error handling JSON marshal. Err: %v", err)
-	}
-
-	w.Write(jsonResp)
-}
-`)
-}
